indented level-based tree drawing algorithm
Level-based tree drawing is a common algorithm that produces intuitive
and clear presentations of hierarchically structured information.
However, new applications often introduces new aesthetic requirements
that call for new tree drawing methods. In this paper, we propose an
indented level-based tree drawing algorithm for visualizing parse trees
of English language. This algorithm displays a tree with an aspect
ratio that fits the aspect ratio of the newer computer displays, while
presenting the words in a way that is easy to read. We discuss the
design of the algorithm and its application in text visualization for
linguistic analysis and language learning. An efficient and practical
implementation of the algorithm is also presented.